+++ src/components/organisms/HomeworksDashboardTeacher.unit.js
@@ -12,9 +12,34 @@ describe("@components/organisms/HomeworksDashboardTeacher", () => {
 		homeworks: {
 			getters: {
 				getLoading: () => false,
+				isListFilled: () => true,
 				getOverDueHomeworks: () => overDueHomeworksTeacher,
 				getOpenHomeworksWithDueDate: () => dueDateHomeworksTeacher,
 				getOpenHomeworksWithoutDueDate: () => noDueDateHomeworksTeacher,
+			},
+		},
+	};
+
+	const mockStoreOnlyWithoutDueDate = {
+		homeworks: {
+			getters: {
+				loading: () => false,
+				isListFilled: () => true,
+				getOverDueHomeworks: () => [],
+				getOpenHomeworksWithDueDate: () => [],
+				getOpenHomeworksWithoutDueDate: () => noDueDateHomeworksTeacher,
+			},
+		},
+	};
+
+	const mockStoreOnlyWithDueDate = {
+		homeworks: {
+			getters: {
+				loading: () => false,
+				isListFilled: () => true,
+				getOverDueHomeworks: () => overDueHomeworksTeacher,
+				getOpenHomeworksWithDueDate: () => dueDateHomeworksTeacher,
+				getOpenHomeworksWithoutDueDate: () => [],
 			},
 		},
 	};
@@ -27,7 +52,7 @@ describe("@components/organisms/HomeworksDashboardTeacher", () => {
 
 	it(...isValidComponent(HomeworksDashboardTeacher));
 
-	it("Should render homeworks list component", () => {
+	it("Should render homeworks list component, with both panels expanded per default", () => {
 		const wrapper = mount(HomeworksDashboardTeacher, {
 			...createComponentMocks({
 				i18n: true,
@@ -37,6 +62,53 @@ describe("@components/organisms/HomeworksDashboardTeacher", () => {
 			vuetify,
 		});
 
+		const expansionPanels = wrapper.findAll(".v-expansion-panel");
+
 		expect(wrapper.findComponent(HomeworksList).exists()).toBe(true);
+		expect(expansionPanels.exists()).toBe(true);
+		expect(expansionPanels.at(0).classes()).toContain(
+			"v-expansion-panel--active"
+		);
+		expect(expansionPanels.at(1).classes()).toContain(
+			"v-expansion-panel--active"
+		);
+	});
+
+	it("Should render only active 'no due date' panel, if the other panel is empty", async () => {
+		const wrapper = mount(HomeworksDashboardTeacher, {
+			...createComponentMocks({
+				i18n: true,
+				vuetify: true,
+				store: mockStoreOnlyWithoutDueDate,
+			}),
+			vuetify,
+		});
+
+		const expansionPanels = wrapper.findAll(".v-expansion-panel");
+
+		expect(expansionPanels.exists()).toBe(true);
+		expect(expansionPanels).toHaveLength(1);
+		expect(expansionPanels.at(0).classes()).toContain(
+			"v-expansion-panel--active"
+		);
+	});
+
+	it("Should render only active 'with due date' panel, if the other panel is empty", async () => {
+		const wrapper = mount(HomeworksDashboardTeacher, {
+			...createComponentMocks({
+				i18n: true,
+				vuetify: true,
+				store: mockStoreOnlyWithDueDate,
+			}),
+			vuetify,
+		});
+
+		const expansionPanels = wrapper.findAll(".v-expansion-panel");
+
+		expect(expansionPanels.exists()).toBe(true);
+		expect(expansionPanels).toHaveLength(1);
+		expect(expansionPanels.at(0).classes()).toContain(
+			"v-expansion-panel--active"
+		);
 	});
 });