+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 ## Unreleased
 
 - SC-9159 - Redirect the routes for news for the new v3 API
+- SC-9189 - Adding expansion panels to teacher's task dashboard
 
 ### Added