+++ CHANGELOG.md
@@ -9,12 +9,10 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 ## Unreleased
 
 - SC-9159 - Redirect the routes for news for the new v3 API
-- SC-9189 - Adding expansion panels to teacher's task dashboard
 
 ### Added
 
 - SC-9090 - implemented a loading state in the service-template
-- SC-9189 - Adding expansion panels to teacher's task dashboard
 
 ### Changed
 
@@ -30,6 +28,12 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 - SC-9093 - Load default language and timezone from server
 - SC-9158 - fixed launch config to debug a single test
 - SC-9160 - fixed getStatus getter in service-template
+
+## [26.5.2] - 2021-07-12
+
+### Added
+
+- SC-9189 - Adding expansion panels to teacher's task dashboard
 
 ## [26.5.1] - 2021-06-30